Add files via upload
Usage: "Ctlr+Shift+G" on the g4 file in the IDE.  This generates a new, hopefully correct, lexer and parser files.  In /gen/, i have a class called POTLUCKRDEMO.java, run that.  Make sure to specify the .txt file.  

I figured out how to printout values with this code: 
plan{recipe{cucumber, onion}} --> check the system.out call in the g4 file.  It currently prints [recipe,cucumber,onion].  I stored it as an ArrayList so far.  We can think of a better way of 'outputting' something, but each may be drastically different.  We need to pass the ArrayList variable upwards to the top start rule I think for 'proper usage'.